projects/02_learn_django_for_professionals/14_search/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap5==0.6
Django==5.1
django-allauth==64.1.0
django-crispy-forms==1.14.0
idna==3.8
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tzdata==2024.1
urllib3==2.2.2
environs[django]==9.5.0
PyJWT==2.7.0
cryptography==41.0.3
pillow==9.0.0